finish implementing #764
- currently untested
