+++ src/hooks/useFetch.ts
@@ -2,13 +2,15 @@
 
 
 import { useState, useEffect } from 'react';
-import axios, { AxiosRequestConfig } from 'axios';
+import axios, { AxiosRequestConfig, CancelTokenSource } from 'axios';
 
 
 
+
+type DataType<T> = T | null;
 
 type FetchResponse<T> = {
-    data: T | null;
+    data: DataType<T>;
     isLoading: boolean;
     isError: boolean;
     error: any;
@@ -22,11 +24,15 @@ const EXTRA_FETCH_LOGS = false;
 
 
 
-export const useFetch = <T = any>(url: string): FetchResponse<T> => {
+export const useFetch = <
+    T = any,
+>(
+    url: string,
+): FetchResponse<T> => {
     const [isLoading, setIsLoading] = useState<boolean>(false);
     const [isError, setIsError] = useState<boolean>(false);
     const [error, setError] = useState<any>(null);
-    const [data, setData] = useState<T | null>(null);
+    const [data, setData] = useState<DataType<T>>(null);
 
     if (EXTRA_FETCH_LOGS) console.log(`useFetch fired with url: ${url}`);
 
@@ -88,7 +94,7 @@ export const useFetch = <T = any>(url: string): FetchResponse<T> => {
                         console.log(`Request canceled for url: ${url}`);
                         return;
                     }
-                    
+
                     console.log(`Error in useFetch: ${error}`);
 
                     setIsError(true);
@@ -111,6 +117,56 @@ export const useFetch = <T = any>(url: string): FetchResponse<T> => {
 };
 
 
-export const fetchWithAxios = async <T = any>(url: string): FetchResponse<T> => {
-    
+
+
+export const fetchWithAxios = async <
+    T = any,
+>(
+    url: string,
+    source: CancelTokenSource,
+): Promise<DataType<T>> => {
+    if (EXTRA_FETCH_LOGS) console.log(`fetchWithAxios fired with url: ${url}`);
+
+
+    if (!url) {
+        if (EXTRA_FETCH_LOGS) console.log("No url provided");
+
+        source.cancel();
+
+        return null;
+    }
+
+
+    try {
+        if (EXTRA_FETCH_LOGS) console.log("fetchWithAxios try block fired");
+
+        const options: AxiosRequestConfig = {
+            url: url,
+            cancelToken: source.token,
+        };
+
+
+        if (EXTRA_FETCH_LOGS) console.log(`call axios with options: ${JSON.stringify(options)}`);
+
+        const axiosResponse = await axios(options);
+
+        if (EXTRA_FETCH_LOGS) console.log(`axiosResponse: ${JSON.stringify(axiosResponse)}`);
+
+        if (axiosResponse.status !== 200) throw `Error: ${axiosResponse.status}`;
+
+
+        const data = axiosResponse.data;
+
+
+        return data;
+    } catch (error) {
+        if (axios.isCancel(error)) {
+            console.log(`Request canceled for url: ${url}`);
+            return null;
+        }
+
+        console.error(`Error in fetchWithAxios: ${error}`);
+
+        return null;
+    }
 };
